Kinect follower now publishes face coordinates.
Face coordinates are published in the form of 32-bit floating point
numbers. The fields in the message are called x and y, mapping to the
vertical plane the kinect is facing into. These coordinates are from the
kinect's point of view and may require transformation to be valid in other
physical locations on the robot. Z coordinates (how far away the user is)
can be published if needed.

Updated the make lists file so that the sonar isn't activated on the
robot, since we're not using it.